+++ kernel/src/main/java/org/kframework/backend/unparser/UnflattenBackend.java
@@ -2,6 +2,7 @@
 package org.kframework.backend.unparser;
 
 import org.kframework.kil.loader.Context;
+import org.kframework.kompile.KompileOptions;
 import org.kframework.utils.Stopwatch;
 import org.kframework.utils.file.FileUtil;
 
@@ -10,7 +11,7 @@ import com.google.inject.Inject;
 public class UnflattenBackend extends UnparserBackend {
 
     @Inject
-    UnflattenBackend(Stopwatch sw, Context context, FileUtil files) {
-        super(sw, context, true, files);
+    UnflattenBackend(Stopwatch sw, Context context, KompileOptions options, FileUtil files) {
+        super(sw, context, options, true, files);
     }
 }