+++ kernel/src/main/java/org/kframework/backend/unparser/UnparserBackend.java
@@ -2,31 +2,33 @@
 package org.kframework.backend.unparser;
 
 import org.apache.commons.io.FilenameUtils;
-import org.kframework.backend.BasicBackend;
+import org.kframework.backend.PosterBackend;
 import org.kframework.kil.Definition;
 import org.kframework.kil.loader.Context;
+import org.kframework.kompile.KompileOptions;
 import org.kframework.krun.ConcretizeSyntax;
 import org.kframework.utils.Stopwatch;
 import org.kframework.utils.file.FileUtil;
 
 import com.google.inject.Inject;
 
-public class UnparserBackend extends BasicBackend {
+public class UnparserBackend extends PosterBackend {
     private boolean unflattenFirst; // unflatten syntax before unparsing
 
     private final FileUtil files;
+    private final KompileOptions options;
 
     @Inject
-    UnparserBackend(Stopwatch sw, Context context, FileUtil files) {
+    UnparserBackend(Stopwatch sw, Context context, KompileOptions options, FileUtil files) {
         super(sw, context);
         this.unflattenFirst = false;
         this.files = files;
+        this.options = options;
     }
 
-    public UnparserBackend(Stopwatch sw, Context context, boolean unflattenFirst, FileUtil files) {
-        super(sw, context);
+    public UnparserBackend(Stopwatch sw, Context context, KompileOptions options, boolean unflattenFirst, FileUtil files) {
+        this(sw, context, options, files);
         this.unflattenFirst = unflattenFirst;
-        this.files = files;
     }
 
     @Override
@@ -41,21 +43,6 @@ public class UnparserBackend extends BasicBackend {
         String unparsedText = unparserFilter.getResult();
 
         files.saveToDefinitionDirectory(FilenameUtils.removeExtension(options.mainDefinitionFile().getName()) + ".unparsed.k", unparsedText);
-    }
-
-    @Override
-    public String getDefaultStep() {
-        return "FirstStep";
-    }
-
-    @Override
-    public boolean documentation() {
-        return false;
-    }
-
-    @Override
-    public boolean generatesDefinition() {
-        return false;
     }
 
 }