+++ kernel/src/main/java/org/kframework/utils/inject/DefinitionLoadingModule.java
@@ -9,6 +9,7 @@ import org.kframework.kil.Definition;
 import org.kframework.kil.loader.Context;
 import org.kframework.kompile.KompileOptions;
 import org.kframework.main.GlobalOptions;
+import org.kframework.main.Tool;
 import org.kframework.utils.BinaryLoader;
 import org.kframework.utils.Stopwatch;
 import org.kframework.utils.errorsystem.KExceptionManager;
@@ -47,9 +48,13 @@ public class DefinitionLoadingModule extends AbstractModule {
     }
 
     @Provides
-    Definition definition(BinaryLoader loader, FileUtil files) {
+    Definition definition(Tool tool, BinaryLoader loader, FileUtil files) {
+        if (tool == Tool.KDOC) {
+            return loader.loadOrDie(Definition.class, files.resolveKompiled("definition-concrete.bin"));
+        }
         return loader.loadOrDie(Definition.class, files.resolveKompiled("definition.bin"));
     }
+
 
     @Provides
     KompileOptions kompileOptions(Context context, FileUtil files) {